add copy button home

+++ lib/widgets/money_repository.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class EarningsRepo {
-  static ValueNotifier<int> earnings = ValueNotifier(0);
+  static ValueNotifier<double> earnings = ValueNotifier(0);
 
   static void increamentEarnings(int value) {
     earnings.value = earnings.value + value;
@@ -10,7 +10,7 @@ class EarningsRepo {
 
   setEarnings() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
-    prefs.setInt('earnings', earnings.value);
+    prefs.setDouble('earnings', earnings.value);
   }
 }
